refactor: Delete types submodule, inline everything into source file
There is no need for complex submodules in the scope of the package going forward,
so we're working to make everything easily accessible in source files.

Next, we'll refactor the return of `parametrize` and `product` to return iterators,
to be able to dump the whole memo construct.
