--- src/nnbench/types/__init__.py
@@ -1,5 +1,0 @@
-"""Custom types defined by ``nnbench`` for benchmarking workloads."""
-
-from .benchmark import Benchmark, BenchmarkRecord, Parameters, State
-from .interface import Interface
-from .memo import Memo, cached_memo
